Add APIServerErrorBudgetBurn alerting name and remove slo_test.go
